Use vscode-task-runner and improve build reliability
[ci skip]

+++ build.py
@@ -8,14 +8,15 @@ from typing import List
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
 DIST_DIR = os.path.join(THIS_DIR, "dist")
 
-PX4_DIR = os.path.join(THIS_DIR, "build", "PX4-Autopilot")
+BUILD_DIR = os.path.join(THIS_DIR, "build")
+PX4_DIR = os.path.join(BUILD_DIR, "PX4-Autopilot")
 
 # warning, v1.10.2 does not appear to build anymore
 with open(os.path.join(THIS_DIR, ".px4-version"), "r") as fp:
     PX4_VERSION = fp.read().strip()
 
 if PX4_VERSION < "v1.13.0":
-    PYMAVLINK_DIR = os.path.join(THIS_DIR, "build", "pymavlink")
+    PYMAVLINK_DIR = os.path.join(BUILD_DIR, "pymavlink")
 else:
     PYMAVLINK_DIR = os.path.join(
         PX4_DIR,
@@ -28,10 +29,25 @@ else:
 
 
 def print2(msg: str) -> None:
+    """
+    Print statement with a prefix.
+    """
     print(f"--- {msg}", flush=True)
 
 
+def touch_file(filename: str) -> None:
+    """
+    Creates the given file with no contet
+    """
+    with open(filename, "w"):
+        pass
+
+
 def clean_directory(directory: str, file_endings: List[str]) -> None:
+    """
+    Clean a directory of files with specific ends of the filenames
+    """
+
     # cancel if the directory is not already there
     if not os.path.isdir(directory):
         return
@@ -42,6 +58,12 @@ def clean_directory(directory: str, file_endings: List[str]) -> None:
 
 
 def clone_pymavlink() -> None:
+    """
+    Clone pymavlink.
+    """
+    if PX4_VERSION >= "v1.13.0":
+        return
+
     if os.path.isdir(PYMAVLINK_DIR):
         # update the checkout if we already have it
         print2("Updating pymavlink")
@@ -56,6 +78,13 @@ def clone_pymavlink() -> None:
 
 
 def clone_px4() -> None:
+    """
+    Clone and patch PX4.
+    """
+
+    # file to record if PX4 has been patched
+    check_patch_file = os.path.join(BUILD_DIR, ".px4-patched")
+
     if os.path.isdir(PX4_DIR):
         # first, figure out what version we have locally
         local_version = next(
@@ -67,18 +96,12 @@ def clone_px4() -> None:
             .splitlines()
             if l.strip().startswith("refs")
         )
+        # if version does not match, nuke it
         if local_version != PX4_VERSION:
             print(f"Existing PX4 checkout is {local_version}, re-cloning")
-            shutil.rmtree(PX4_DIR)
+            shutil.rmtree(BUILD_DIR)
             clone_px4()
 
-        # reset checkout if we already have it
-        # this will fail on PX4 version changes
-        print2("Resetting PX4 checkout")
-        subprocess.check_call(["git", "fetch", "origin"], cwd=PX4_DIR)
-        subprocess.check_call(["git", "checkout", PX4_VERSION], cwd=PX4_DIR)
-        subprocess.check_call(["git", "reset", "--hard", PX4_VERSION], cwd=PX4_DIR)
-        subprocess.check_call(["git", "pull", "--recurse-submodules"], cwd=PX4_DIR)
     else:
         # clone fresh
         print2("Cloning PX4")
@@ -96,16 +119,42 @@ def clone_px4() -> None:
             ]
         )
 
-    print2("Applying PX4 patch")
+    if not os.path.isfile(check_patch_file):
+        print2("Applying PX4 patch")
+        subprocess.check_call(
+            [
+                "git",
+                "apply",
+                "--ignore-space-change",
+                "--ignore-whitespace",
+                os.path.join(
+                    THIS_DIR, "patches", f"hil_gps_heading_{PX4_VERSION}.patch"
+                ),
+            ],
+            cwd=PX4_DIR,
+        )
+
+        # record that it has been patched
+        touch_file(check_patch_file)
+
+
+def install_dependencies() -> None:
+    """
+    Install any needed dependencies
+    """
+    print2("Installing Python dependencies")
+    subprocess.check_call(
+        [sys.executable, "-m", "pip", "install", "--upgrade", "pip", "wheel"]
+    )
     subprocess.check_call(
         [
-            "git",
-            "apply",
-            "--ignore-space-change",
-            "--ignore-whitespace",
-            os.path.join(THIS_DIR, "patches", f"hil_gps_heading_{PX4_VERSION}.patch"),
-        ],
-        cwd=PX4_DIR,
+            sys.executable,
+            "-m",
+            "pip",
+            "install",
+            "-r",
+            os.path.join(PYMAVLINK_DIR, "requirements.txt"),
+        ]
     )
 
 
@@ -114,20 +163,27 @@ def build_pymavlink(
 ) -> None:
     print2("Generating pymavlink package")
 
-    print2("Applying Pymavlink patch")
-    subprocess.check_call(["git", "reset", "--hard"], cwd=PYMAVLINK_DIR)
-    subprocess.check_call(
-        [
-            "git",
-            "apply",
-            "--ignore-space-change",
-            "--ignore-whitespace",
-            os.path.join(THIS_DIR, "patches", f"pymavlink_{PX4_VERSION}.patch"),
-        ],
-        cwd=PYMAVLINK_DIR,
-    )
+    # file to record if Pymavlink has been patched
+    check_patch_file = os.path.join(BUILD_DIR, ".pymavlink-patched")
+
+    if not os.path.isfile(check_patch_file):
+        print2("Applying Pymavlink patch")
+        subprocess.check_call(["git", "reset", "--hard"], cwd=PYMAVLINK_DIR)
+        subprocess.check_call(
+            [
+                "git",
+                "apply",
+                "--ignore-space-change",
+                "--ignore-whitespace",
+                os.path.join(THIS_DIR, "patches", f"pymavlink_{PX4_VERSION}.patch"),
+            ],
+            cwd=PYMAVLINK_DIR,
+        )
+        # record that it has been patched
+        touch_file(check_patch_file)
 
     # copy message definitions from px4 so we're using the exact same version
+    print2("Copying message definitions")
     shutil.rmtree(
         os.path.join(PYMAVLINK_DIR, "message_definitions", "v1.0"),
         ignore_errors=True,
@@ -140,14 +196,21 @@ def build_pymavlink(
     pymavlink_dist_dir = os.path.join(PYMAVLINK_DIR, "dist")
 
     # clean the pymavlink build and target dirs
-    clean_directory(pymavlink_dist_dir, [".tar.gz", ".whl"])
-    clean_directory(DIST_DIR, [".tar.gz", ".whl"])
+    # print2("Cleaning output")
+    # clean_directory(pymavlink_dist_dir, [".tar.gz", ".whl"])
+    # clean_directory(DIST_DIR, [".tar.gz", ".whl"])
 
     # make a new environment with the mavlink dialect set
     new_env = os.environ.copy()
     new_env["MAVLINK_DIALECT"] = "bell"
+    print2("Building package")
     subprocess.check_call(
-        [sys.executable, "setup.py", "sdist"],  # , "bdist_wheel",],
+        [
+            sys.executable,
+            "setup.py",
+            "sdist",
+            "bdist_wheel",
+        ],
         cwd=PYMAVLINK_DIR,
         env=new_env,
     )
@@ -162,6 +225,7 @@ def build_pymavlink(
     # generate lua plugins for Wireshark
     # https://mavlink.io/en/guide/wireshark.html
     if should_build_wireshark:
+        print2("Building wireshark plugin")
         subprocess.check_call(
             [
                 sys.executable,
@@ -182,8 +246,8 @@ def build_px4(targets: List[str], version: str) -> None:
     px4_build_dir = os.path.join(PX4_DIR, "build")
 
     # clean the PX4 build and target dir
-    clean_directory(px4_build_dir, [".px4"])
-    clean_directory(DIST_DIR, [".px4"])
+    # clean_directory(px4_build_dir, [".px4"])
+    # clean_directory(DIST_DIR, [".px4"])
 
     for target in targets:
         subprocess.check_call(["make", target, "-j"], cwd=PX4_DIR)
@@ -202,27 +266,14 @@ def main(
 ) -> None:
     os.makedirs(DIST_DIR, exist_ok=True)
 
-    if PX4_VERSION < "v1.13.0":
-        clone_pymavlink()
+    # clone pymavlink if necessary
+    clone_pymavlink()
 
     # get px4 cloned
     clone_px4()
 
     # install python dependencies for pymavlink
-    print2("Installing Python dependencies")
-    subprocess.check_call(
-        [sys.executable, "-m", "pip", "install", "--upgrade", "pip", "wheel"]
-    )
-    subprocess.check_call(
-        [
-            sys.executable,
-            "-m",
-            "pip",
-            "install",
-            "-r",
-            os.path.join(PYMAVLINK_DIR, "requirements.txt"),
-        ]
-    )
+    install_dependencies()
 
     # build directory paths
     if PX4_VERSION < "v1.13.0":
@@ -249,27 +300,30 @@ def main(
 
     bell_xml_def = os.path.join(message_definitions_dir, "bell.xml")
 
-    print2("Injecting Bell MAVLink message")
-    shutil.copyfile(os.path.join(THIS_DIR, "bell.xml"), bell_xml_def)
+    # file to record if mavlink has been committed
+    check_patch_file = os.path.join(BUILD_DIR, ".mavlink-commited")
 
-    # generate the mavlink C code
-    if PX4_VERSION < "v1.13.0":
-        subprocess.check_call(
-            [
-                sys.executable,
-                "-m",
-                "pymavlink.tools.mavgen",
-                "--lang=C",
-                "--wire-protocol=2.0",
-                f"--output={generated_message_dir}",
-                bell_xml_def,
-            ],
-            cwd=os.path.join(PYMAVLINK_DIR, ".."),
-        )
+    if not os.path.isfile(check_patch_file):
+        print2("Injecting Bell MAVLink message")
+        shutil.copyfile(os.path.join(THIS_DIR, "bell.xml"), bell_xml_def)
 
-    # git config does not matter, just need *something* to commit,
-    # they're not pushed anywhere
-    if subprocess.run(["git", "config", "user.name"]).returncode != 0:
+        # generate the mavlink C code
+        if PX4_VERSION < "v1.13.0":
+            subprocess.check_call(
+                [
+                    sys.executable,
+                    "-m",
+                    "pymavlink.tools.mavgen",
+                    "--lang=C",
+                    "--wire-protocol=2.0",
+                    f"--output={generated_message_dir}",
+                    bell_xml_def,
+                ],
+                cwd=os.path.join(PYMAVLINK_DIR, ".."),
+            )
+
+        # git config does not matter, just need *something* to commit,
+        # they're not pushed anywhere
         subprocess.check_call(
             ["git", "config", "user.email", "github-bot@bellflight.com"], cwd=PX4_DIR
         )
@@ -277,18 +331,21 @@ def main(
             ["git", "config", "user.name", "Github Actions"], cwd=PX4_DIR
         )
 
-    # changes need to be committed to build
-    subprocess.check_call(["git", "add", "."], cwd=PX4_DIR)
-    subprocess.check_call(
-        [
-            "git",
-            "commit",
-            "--no-gpg-sign",
-            "-m",
-            "Local commit to facilitate build",
-        ],
-        cwd=PX4_DIR,
-    )
+        # changes need to be committed to build
+        subprocess.check_call(["git", "add", "."], cwd=PX4_DIR)
+        subprocess.check_call(
+            [
+                "git",
+                "commit",
+                "--no-gpg-sign",
+                "-m",
+                "Local commit to facilitate build",
+            ],
+            cwd=PX4_DIR,
+        )
+
+        # record that it has been committed
+        touch_file(check_patch_file)
 
     if should_build_pymavlink:
         build_pymavlink(message_definitions_dir, bell_xml_def, should_build_wireshark)
